ss
ss

+++ js/almacenamiento.js
@@ -266,12 +266,14 @@ fechaderegistro: null,
 									eliminarregExtintores: function(tx){
 									 for(i = 0; i<almacen.myArray.length; i++) 
 										    {
-										    	if((almacen.myArray[i] != "") && (almacen.myArray[i] != undefined))
+										    	if((almacen.myArray[i] != "") && (almacen.myArray[i] != undefined) && (almacen.myArray[i] != "null"))
 										    	{
 										    		tx.executeSql("DELETE FROM ita_sh_reg_ext where id_ext = '"+almacen.myArray[i]+"'" );
 													navigator.notification.alert("DELETE FROM ita_sh_reg_ext where id_ext = '"+almacen.myArray[i] +"'", null, "Correcto", "Aceptar");
     											}
-        									}      
+        									} 
+
+        									almacen.myArray.length = 0;    
 
 	},
 	/*ALMACENAMIENTO PARA HIDRANTES*/
